# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Riviera Building Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /units:
    get:
      operationId: listUnits
      tags: [units]
      responses:
        "200":
          description: List of units
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Unit"
    post:
      operationId: createUnit
      tags: [units]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
  /units/{id}:
    get:
      operationId: getUnit
      tags: [units]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
        "404":
          description: Not found
    patch:
      operationId: updateUnit
      tags: [units]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
    delete:
      operationId: deleteUnit
      tags: [units]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Tenants ────────────────────────────────────────────────────────────────
  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      responses:
        "200":
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
    post:
      operationId: createTenant
      tags: [tenants]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /tenants/{id}:
    get:
      operationId: getTenant
      tags: [tenants]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Not found
    patch:
      operationId: updateTenant
      tags: [tenants]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
    delete:
      operationId: deleteTenant
      tags: [tenants]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Contracts ──────────────────────────────────────────────────────────────
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    patch:
      operationId: updateSettings
      tags: [settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
  /settings/exchange-rates:
    get:
      operationId: getExchangeRates
      tags: [settings]
      responses:
        "200":
          description: Current exchange rates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRates"
    patch:
      operationId: updateExchangeRates
      tags: [settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExchangeRatesUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRates"
  /settings/users:
    get:
      operationId: listUsers
      tags: [settings]
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /settings/users/{id}:
    patch:
      operationId: updateUser
      tags: [settings]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [settings]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    User:
      type: object
      required: [id, username, name, role]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, manager, viewer]
        createdAt:
          type: string
    UserInput:
      type: object
      required: [username, name, password, role]
      properties:
        username:
          type: string
        name:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, manager, viewer]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        password:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, manager, viewer]

    # ─── Dashboard ─────────────────────────────────────────────────────────────
...
    Unit:
      type: object
      required: [id, unitNumber, floor, type, area, status]
      properties:
        id:
          type: integer
        unitNumber:
          type: string
        floor:
          type: string
        type:
          type: string
          enum: [office, shop, warehouse, apartment, other]
        area:
          type: number
        status:
          type: string
          enum: [vacant, occupied, maintenance]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    UnitInput:
      type: object
      required: [unitNumber, floor, type, area, status]
      properties:
        unitNumber:
          type: string
        floor:
          type: string
        type:
          type: string
          enum: [office, shop, warehouse, apartment, other]
        area:
          type: number
        status:
          type: string
          enum: [vacant, occupied, maintenance]
        description:
          type: ["string", "null"]
    UnitUpdate:
      type: object
      properties:
        unitNumber:
          type: string
        floor:
          type: string
        type:
          type: string
          enum: [office, shop, warehouse, apartment, other]
        area:
          type: number
        status:
          type: string
          enum: [vacant, occupied, maintenance]
        description:
          type: ["string", "null"]

    # ─── Tenants ───────────────────────────────────────────────────────────────
    Tenant:
      type: object
      required: [id, name, type, phone]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [individual, company]
        phone:
          type: string
        email:
          type: ["string", "null"]
        idNumber:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        balance:
          type: number
        createdAt:
          type: string
    TenantInput:
      type: object
      required: [name, type, phone]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [individual, company]
        phone:
          type: string
        email:
          type: ["string", "null"]
        idNumber:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    TenantUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [individual, company]
        phone:
          type: string
        email:
          type: ["string", "null"]
        idNumber:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # ─── Contracts ─────────────────────────────────────────────────────────────
...
    Settings:
      type: object
      required: [buildingName, buildingAddress, defaultCurrency]
      properties:
        buildingName:
          type: string
        buildingAddress:
          type: string
        defaultCurrency:
          type: string
          enum: [ILS, USD, JOD]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        taxNumber:
          type: ["string", "null"]
    SettingsUpdate:
      type: object
      properties:
        buildingName:
          type: string
        buildingAddress:
          type: string
        defaultCurrency:
          type: string
          enum: [ILS, USD, JOD]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        taxNumber:
          type: ["string", "null"]
    ExchangeRates:
      type: object
      required: [usdToILS, jodToILS, updatedAt]
      properties:
        usdToILS:
          type: number
        jodToILS:
          type: number
        updatedAt:
          type: string
    ExchangeRatesUpdate:
      type: object
      required: [usdToILS, jodToILS]
      properties:
        usdToILS:
          type: number
        jodToILS:
          type: number